Converted to a single project, enabled Razor Class library, also added a .editorconfig to reduce stlying issues in future :D
