--- src/HC.PageNotFoundManager.Core/Backoffice/PageNotFoundDialogControllerFile.cs
@@ -1,12 +1,0 @@
-﻿using Umbraco.Cms.Core.WebAssets;
-
-namespace HC.PageNotFoundManager.Core.Backoffice
-{
-    public class PageNotFoundDialogControllerFile : JavaScriptFile
-    {
-        public PageNotFoundDialogControllerFile() : base("/App_Plugins/HC.PageNotFound/js/dialog.controller.js")
-        {
-
-        }
-    }
-}
