--- src/HC.PageNotFoundManager.Core/Backoffice/PageNotFoundResourceFile.cs
@@ -1,12 +1,0 @@
-﻿using Umbraco.Cms.Core.WebAssets;
-
-namespace HC.PageNotFoundManager.Core.Backoffice
-{
-    public class PageNotFoundResourceFile : JavaScriptFile
-    {
-        public PageNotFoundResourceFile() : base("/App_Plugins/HC.PageNotFound/js/resource.js")
-        {
-
-        }
-    }
-}
