--- src/HC.PageNotFoundManager.Core/Config/PageNotFound.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace HC.PageNotFoundManager.Core.Config
-{
-    /// <summary>
-    /// Class to represent the settings from appsettings
-    /// </summary>
-    public class PageNotFound
-    {
-        public PageNotFound()
-        {
-            UserGroups = Array.Empty<string>();
-        }
-        public string[] UserGroups { get; set; }
-    }
-}
