--- src/HC.PageNotFoundManager.Core/Constants/Constants.cs
@@ -1,11 +1,0 @@
-﻿namespace HC.PageNotFoundManager.Core
-{
-    public static class Constants
-    {
-        public const string Base = "HCS";
-        public const string BackOffice = Base + "PageNotFound";
-
-        public const string MenuAlias = "pageNotFoundManagement";
-        public const string MenuLabel = "Manage 404 Page";
-    }
-}
