--- src/HC.PageNotFoundManager.Core/Models/PageNotFoundRequest.cs
@@ -1,10 +1,0 @@
-﻿namespace HC.PageNotFoundManager.Core.Models
-{
-    public class PageNotFoundRequest
-    {
-        public int ParentId { get; set; }
-
-        public int NotFoundPageId { get; set; }
-
-    }
-}
