--- src/HC.PageNotFoundManager.Core/Startup/StartupComposer.cs
@@ -1,13 +1,0 @@
-﻿using Umbraco.Cms.Core.Composing;
-using Umbraco.Cms.Core.DependencyInjection;
-
-namespace HC.PageNotFoundManager.Core.Startup
-{    
-    public class StartupComposer : IComposer
-    {
-        public void Compose(IUmbracoBuilder builder)
-        {
-            builder.UsePageNotFoundManager();
-        }
-    }
-}
