+++ src/HC.PageNotFoundManager/Constants/Constants.cs
@@ -1,0 +1,13 @@
+﻿namespace HC.PageNotFoundManager.Constants
+{
+    public static class Constants
+    {
+        public const string BackOffice = Base + "PageNotFound";
+
+        public const string Base = "HCS";
+
+        public const string MenuAlias = "pageNotFoundManagement";
+
+        public const string MenuLabel = "Manage 404 Page";
+    }
+}
