+++ src/HC.PageNotFoundManager/Models/PageNotFoundRequest.cs
@@ -1,0 +1,10 @@
+﻿namespace HC.PageNotFoundManager.Models
+{
+    public class PageNotFoundRequest
+    {
+
+        public int NotFoundPageId { get; set; }
+
+        public int ParentId { get; set; }
+    }
+}
